update stem len type to uint32

+++ topic.go
@@ -310,11 +310,11 @@ func (t *Topic) Publish(ctx context.Context, data []byte, opts ...PubOpt) error 
 		}
 	}
 
-	stemLenDiff, err := rand.Int(rand.Reader, big.NewInt(GossipSubStemMax-GossipSubStemMin))
+	stemLenDiff, err := rand.Int(rand.Reader, big.NewInt(int64(GossipSubStemMax-GossipSubStemMin)))
 	if err != nil {
 		return err
 	}
-	stemLen := GossipSubStemMin + stemLenDiff.Uint64()
+	stemLen := GossipSubStemMin + uint32(stemLenDiff.Uint64())
 	m.Stem = &stemLen
 	return t.p.val.PushLocal(&Message{m, "", t.p.host.ID(), nil, pub.local})
 }
